Enable Travis CI
Add .travis.yml and add the build status badge to the README.

Example build on my fork: https://travis-ci.org/rfjakob/cryptoWallet/builds/394753550

You will need to enable Travis on your repository on
https://travis-ci.org/profile/xaionaro-go so that automatic builds
and the status badge works.

The Go versions that are built are the same that gocryptfs uses
in its travis config.